Revert branch to state of commit eb797fe
