+++ packages/eslint-plugin-slds/src/rules/v9/enforce-component-hook-naming-convention.ts
@@ -1,8 +1,6 @@
 import { Rule } from 'eslint';
 import metadata from '@salesforce-ux/sds-metadata';
 import ruleMessages from '../../config/rule-messages.yml';
-import { findCssVariables } from '../../utils/css-var-utils';
-import { reportDeprecatedHook, reportWithoutFix } from '../../utils/eslint-report-utils';
 
 const ruleConfig = ruleMessages['enforce-component-hook-naming-convention'];
 const { type, description, url, messages } = ruleConfig;
@@ -34,73 +32,56 @@ export default {
   create(context) {
     return {
       /*
-       * Handle component hooks in CSS declarations (property names)
-       * Example: --slds-c-accordion-color-border: #000;
+       * Handle component hooks in CSS declarations
+       * Limitation: For example:
+       * .testClass{
+       *    --slds-c-accordion-section-color-background: var(--slds-c-accordion-section-color-border);
+       *  }
+       * var in value is not detected, because eslint treats above statement as custom property assignment 
+       * and value treated as raw string, not parsed into function nodes
        */
-      "Declaration[property=/^--slds-c-/]"(node) {
-        const hookName = node.property;
+      "Declaration[property=/^--slds-c-/], Function[name='var'] Identifier[name=/^--slds-c-/]"(node) {
+        let hookName: string;
+        let reportNode = node;
         
-        // Check if this property itself is deprecated
-        if (!shouldIgnoreDetection(hookName)) {
-          const suggestedMatch = slds1DeprecatedComponentHooks[hookName];
-          reportDeprecatedHook(context, node, hookName, suggestedMatch);
+        if (node.type === "Declaration") {
+          // CSS custom property declaration: --slds-c-...
+          hookName = node.property;
+        } else if (node.type === "Identifier") {
+          // Inside var() function: var(--slds-c-...)
+          hookName = node.name;
+        } else {
+          return;
         }
-        
-        // Note: We don't return early here because we also need to check the value
-        // for other deprecated hooks. The broader Declaration selector will handle that.
-      },
 
-      /*
-       * Handle component hooks in var() functions (direct usage)
-       * Example: var(--slds-c-accordion-color-border)
-       */
-      "Function[name='var'] Identifier[name=/^--slds-c-/]"(node) {
-        const hookName = node.name;
-        
         // Skip if hook should be ignored
         if (shouldIgnoreDetection(hookName)) {
           return;
         }
 
         const suggestedMatch = slds1DeprecatedComponentHooks[hookName];
-        reportDeprecatedHook(context, node, hookName, suggestedMatch);
-      },
-
-      /*
-       * Handle component hooks in CSS property values (covering the limitation)
-       * This catches deprecated hooks used as values that aren't caught by the var() function selector
-       * Example: --custom-prop: --slds-c-accordion-color-border; (without var() wrapper)
-       * Example: --slds-c-prop: var(--slds-c-accordion-color-border); (with var() wrapper)
-       */
-      "Declaration"(node) {
-        // Get the raw text of the entire declaration
-        const declarationText = context.sourceCode.getText(node);
         
-        // Find all component hooks in the declaration value (excluding the property name)
-        const allHookMatches = findCssVariables(declarationText, /--slds-c-[a-zA-Z0-9-]+/g);
-        
-        if (allHookMatches.length > 0) {
-          // Process hooks that are NOT the property name itself
-          const valueHooks = allHookMatches.filter(hook => {
-            // Skip the property name itself (already handled by the specific selector)
-            return node.property !== hook;
-          });
-          
-          valueHooks.forEach((hookName: string) => {
-            // Skip if hook should be ignored
-            if (shouldIgnoreDetection(hookName)) {
-              return;
+        context.report({
+          node: reportNode,
+          messageId: 'replace',
+          data: { 
+            oldValue: hookName,
+            suggestedMatch: suggestedMatch
+          },
+          fix(fixer) {
+            if (node.type === "Declaration") {
+              // Replace the property name in CSS declaration
+              const originalText = context.sourceCode.getText(node);
+              const colonIndex = originalText.indexOf(':');
+              const valuePartWithColon = originalText.substring(colonIndex);
+              return fixer.replaceText(node, `${suggestedMatch}${valuePartWithColon}`);
+            } else if (node.type === "Identifier") {
+              // Replace the identifier name in var() function
+              return fixer.replaceText(node, suggestedMatch);
             }
-
-            const suggestedMatch = slds1DeprecatedComponentHooks[hookName];
-            
-            // Report the issue but without auto-fix for complex value parsing
-            reportWithoutFix(context, node, 'replace', {
-              oldValue: hookName,
-              suggestedMatch: suggestedMatch
-            });
-          });
-        }
+            return null;
+          }
+        });
       }
     };
   },
